+++ src/ot/sphere.c
@@ -59,7 +59,7 @@ register CUBE  *cu;
 {
 	FVECT  v1;
 	double  d1, d2;
-	register double  *fa;
+	register FLOAT  *fa;
 	register int  i;
 #define  cent		fa
 #define  rad		fa[3]
